depreceated model nb and divided functionalities between model base, glm and glmnb
also improved importing and exporting to and from submodules
