+++ batchglm/api/models/nb_glm.py
@@ -1,4 +1,2 @@
-from batchglm.models.nb_glm.input import InputData
-from batchglm.models.nb_glm.model import Model
-from batchglm.models.nb_glm.simulator import Simulator
-from batchglm.train.tf.nb_glm.estimator import Estimator
+from batchglm.models.nb_glm import InputData_NBGLM, Model, Simulator
+from batchglm.train.tf.nb_glm import Estimator
