+++ batchglm/models/external.py
@@ -1,1 +1,2 @@
 import batchglm.pkg_constants as pkg_constants
+import batchglm.data as data_utils
