+++ batchglm/models/glm/__init__.py
@@ -1,3 +1,4 @@
 from .base import BasicGLM
+from .input import InputData_GLM, INPUT_DATA_PARAMS
 from .utils import parse_design
 from .utils import closedform_glm_mean, closedform_glm_var
