+++ batchglm/models/glm/external.py
@@ -1,2 +1,4 @@
+from batchglm.models import BasicInputData
+
 from batchglm.utils.linalg import groupwise_solve_lm
 from batchglm.utils.numeric import weighted_mean, weighted_variance
