+++ batchglm/models/nb_glm/__init__.py
@@ -1,2 +1,4 @@
 from .model import Model, XArrayModel
-from .input import InputData
+from .input import InputData_NBGLM
+from .simulator import Simulator
+from .estimator import AbstractEstimator, XArrayEstimatorStore
