+++ batchglm/train/tf/nb_glm/__init__.py
@@ -1,0 +1,1 @@
+from .estimator import Estimator
