+++ batchglm/train/tf/nb_glm/external.py
@@ -1,13 +1,11 @@
 import batchglm.data as data_utils
 
-from batchglm.models.nb_glm.estimator import AbstractEstimator, XArrayEstimatorStore
-from batchglm.models.nb_glm.input import InputData
-from batchglm.models.nb_glm.model import Model
+from batchglm.models.nb_glm import AbstractEstimator, XArrayEstimatorStore, InputData_NBGLM, Model
+import batchglm.models.nb_glm.utils as nb_glm_utils
 
 import batchglm.train.tf.ops as op_utils
 import batchglm.train.tf.train as train_utils
 from batchglm.train.tf.base import TFEstimatorGraph, MonitoredTFEstimator
 
-import batchglm.models.nb_glm.utils as nb_glm_utils
 import batchglm.utils.random as rand_utils
 from batchglm import pkg_constants
