+++ batchglm/train/tf/nb_glm/model.py
@@ -1,9 +1,6 @@
 import logging
-# import pprint
-# from enum import Enum
 
 import tensorflow as tf
-# import tensorflow_probability as tfp
 
 import numpy as np
